First Commit for Alex's changes.
WARNINGS
-This has only been tested on a 100/250 VESC.
-In conjunction with the below changes, there are particular setups
that must be adhered to within the Motor and Balance App configs.
I am planning on making a tutorial to show the specific changes
required.

*IMU Changes*
-Changed accelerometer and gyro full scale ranges
-Added first attempt of a recursive estimator (Kalman filter) instead
of existing biquad filters

*Balance App Changes*
- Added hightemp tiltback for motor and FET's
-fixed derivative terms; trying two methods
-added Kalman filtering on derivative terms
-Work In Progress on non-linear output; it is functioning but
need to change the scaling on the gains to make it
user friendly
-Work In Progress on ADC filtering; added low pass filters
to ADC values, trying to make better functioning footpad sensor
-Work In Progress on nose angle with respect to speed; had this working
and the idea is gain scheduling the PID gains wrt speed. There are currently
some other methods being used within the balance app
that might acheive similar results.