Added non-linear function to the current output. Braking and Acceleration can be tuned separately. I need to follow up with documentation on how to tune through app. Right now, negative values give greater rate of change relative to input. Positive values give lower rate of change relative to input. My testing setup used a known good tune with linear PID terms. Then, apply the non-linear tuning to the gains. I have found positive terms for acceleration and negative terms for braking to feel the best. This gives more modulation for acceleration and less for braking, but you never tail drag.
Additional functionality that was added was a first pass at
using the nunchuk via the NRF dongle (I used a trampa "magic wand")
to adjust the nose angle on the fly. In the setup I have, if you
release the button, it always returns to the base setting. If you don't
have a wand, it assumes the base setting is zero.

I need to follow up with documentation on this.