8-25-2023: Reviewed and tested methods for filtering the derivative term. Tested Mahony and Madgwick Fusion. IMO, Madgwick Fusion is way better than Mahony if you want an aggressive board. All my testing ended up being with Madgwick Fusion after dialing in the response of it.
I am using 1600Hz, low filter, 0 Confidence Decay, and Beta of 0.8.
I would like to use a Beta of 1 or maybe up to 1.5 to get even better
performance. 0.8 seems to strike a balance between response and noise.
Changing the beta value
has a significant effect on the PID gains in the balance app.

This was done by adding different variables as required to the
"app_balance_experiment" portion of the code so the real-time
signals could be viewed in VESC tool. This made it pretty clear from
a time domain perspective what was going on.
Digging deeper into the root of the problem, I also pulled some .CSV
and used a MATLAB script to look at the data in the frequency domain.

The below were major things tested.

-added a pitch buffer
-Simple moving average filter : This helped somewhat but
still was susceptible to ringing
-Exponential moving average filter: Same as above, not much different
-First order low pass filter: There is an implementation already of
a low pass filter, and while it does smooth data as seen in the time
domain, it creates ringing on the signal which limits
the stability margins severly of the PID gains in the ballance app
-Higher order low pass filters: Same as above, and same for the already
implemented biquad filter
- Savitzky–Golay filter: I could not get this to scale correctly
 and abandoned it. Its possible I was missing something simple.

-After testing various arangements of pre-filtering the pitch angle
coming from AHRS, post-filtering the derivative, or both with the
above filters, all of the above did not give the stability margins
I was hoping for.

-I implemented about the simplest first-order discrete Kalman filter,
which actually did help significantly. However, I still could not
get the Balance app gains where I wanted them in terms of performance.
If I set low-ish peak currents, say 80 or 100 amps, I could make a board
that felt GT-ish, but still didn't feel as good as I hoped. For example,
the filter gains had to be turned so low that the high frequency components
were gone, and the derivative term wouldn't act on fast movements. It felt
okay if slow or moderate changes to the pitch angle were occuring, which
seemed fine for chil pavement cruising. I then found through testing
that turning the integral gains down on the FOC PI control would dramatically
improve the stability margins, and I could push higher PID gains within
the balance app, but because the current gains were so much lower to get
there, it was possible to lose control of the motor say under really hard
braking, which was a little dangerous.

After a bit more reading, I decided I could either try and implement a
more complex Kalman filter, or try a few more configurations with
cascading Kalman filters. Since I already had a working implementation,
I tested the Kalman filter on the pitch angle signal, with another Kalman filter
on the Derivative term. After watching the real-time signals of the raw pitch
data vs. the filtered, I found gains for the pre and post filters that seemed
to work pretty well. After testing this, I was able to dramatically increase the
FOC PI gains, and also increase the balance app gains! Finally, this felt
like the best version of the board to date. I want to post some app and
motor configs I am using as well as an updated map of the variables
I have repurposed within VESC tool.

Additional changes were:
balance app:
-added dynamic deadband
-changed the constant tiltback to start with a tiltback angle at zero
speed, and level out at some prescriped motor speed.
BMI_160 wrapper:
-changed ranges of accelerometer to improve signal to noise
ratio based on BMI 160 datasheet
-Additionally, I tried adding the time variable to the data structure,
but gave up on it for now.
The data flow from the IMU to the AHRS library, the drivers, and
to the balance app seems to impart issues in timing, but it would be
a pretty big effort to revise it based on how it is written. I think it is
great for most general purpose app building, but if we wanted
a purely custom Balance Board app, it might be better to do
something different. I was thinking the IMU code could just
live in the balance app, but not sure if that would be best.

It seems like it would be nice/best to get
the timestamp out of the BMI with the accel and gyro data, and use
that to calculate the derivative instead of a separate clock for better
accuracy in the derivative.
-changed pairing of frequency of gyro to accelerometers.
-IMU.c
-added kalman filtering here instead of biquad. This actually doesn't
work very well. I think this is because of the timing between the actual
IMU data and the balance app.